refactor : destory account
