+++ src/components/MyPage/MyPageLayout/index.tsx
@@ -1,13 +1,17 @@
-import React from 'react';
+import React, { useEffect, useCallback } from 'react';
 import PageMenu from '../PageMenu';
 import UserCard from '../UserCard';
 import Modal from 'components/ui/Modal';
 import Loading from 'components/ui/Loading';
 import ConfirmPopUp from 'components/ui/ConfirmPopUp';
 import { Title } from 'atoms';
-import { useDestroyAccount } from 'hooks';
+import { useModal, useApiFetch } from 'hooks';
+import { REQUEST, SUCCESS } from 'hooks/common/useApiFetch';
 import { UserType } from 'types/User';
 import { DESTROY_ACCOUNT_CAPTION } from 'common/constant/string';
+import { destroyAccount } from 'api/user';
+import routes from 'common/constant/routes';
+import Router from 'next/router';
 import * as menus from 'common/constant/mypageDatas';
 import * as S from './style';
 
@@ -22,7 +26,19 @@ const MyPage = ({
   pageName,
   children,
 }: Props): React.ReactElement => {
-  const data = useDestroyAccount(userInfo.userId);
+  const [showModal, modalHandler] = useModal(false);
+  const [destroyResult, destroyDispatch] = useApiFetch(destroyAccount);
+
+  useEffect(() => {
+    if (destroyResult.type === SUCCESS) {
+      modalHandler();
+      Router.push(routes.HOME);
+    }
+  }, [destroyResult.type]);
+
+  const DestroyHandler = useCallback(() => {
+    destroyDispatch({ type: REQUEST, params: [userInfo.userId] });
+  }, []);
 
   return (
     <S.Container>
@@ -37,7 +53,7 @@ const MyPage = ({
           <Loading />
         )}
         <PageMenu datas={menus.GeneralMenu} />
-        <PageMenu datas={menus.DestoryMenu} onClick={data.modalHandler} />
+        <PageMenu datas={menus.DestoryMenu} onClick={modalHandler} />
       </S.SideContainer>
       <S.MainContainer>
         <S.TitleContainer>
@@ -45,11 +61,11 @@ const MyPage = ({
         </S.TitleContainer>
         {children}
       </S.MainContainer>
-      <Modal showModal={data.showModal} modalHandler={data.modalHandler}>
+      <Modal showModal={showModal} modalHandler={modalHandler}>
         <ConfirmPopUp
           contents={DESTROY_ACCOUNT_CAPTION}
-          closeHandler={data.modalHandler}
-          submitHandler={data.DestroyHandler}
+          closeHandler={modalHandler}
+          submitHandler={DestroyHandler}
         />
       </Modal>
     </S.Container>
